refactor: capitalize and pluralize action types file
