+++ src/actions/cms/articles/update.ts
@@ -4,7 +4,7 @@ import prisma from '@/prisma'
 import errorHandler from '@/prisma/errorHandler'
 import { addArticleSectionPart } from '@/cms/articleSections/update'
 import { z } from 'zod'
-import type { ActionReturn } from '@/actions/type'
+import type { ActionReturn } from '@/actions/Types'
 import type { ArticleSection } from '@prisma/client'
 import type { Part } from '@/cms/articleSections/update'
 import type { ReturnType } from './ReturnType'
