+++ src/actions/cms/images/create.ts
@@ -2,7 +2,7 @@
 import prisma from '@/prisma'
 import errorHandler from '@/prisma/errorHandler'
 import type { Image, CmsImage } from '@prisma/client'
-import type { ActionReturn } from '@/actions/type'
+import type { ActionReturn } from '@/actions/Types'
 
 export async function createCmsImage(name: string): Promise<ActionReturn<CmsImage & {image: Image | null}>> {
     try {
