+++ src/actions/cms/links/create.ts
@@ -1,7 +1,7 @@
 'use server'
 import errorHandler from '@/prisma/errorHandler'
 import prisma from '@/prisma'
-import type { ActionReturn } from '@/actions/type'
+import type { ActionReturn } from '@/actions/Types'
 import type { CmsLink } from '@prisma/client'
 
 export async function createCmsLink(name: string): Promise<ActionReturn<CmsLink>> {
