+++ src/actions/images/collections/destroy.ts
@@ -2,7 +2,7 @@
 import prisma from '@/prisma'
 import errorHandler from '@/prisma/errorHandler'
 import type { ImageCollection } from '@prisma/client'
-import type { ActionReturn } from '@/actions/type'
+import type { ActionReturn } from '@/actions/Types'
 
 export async function destroyImageCollection(collectionId: number): Promise<ActionReturn<ImageCollection>> {
     try {
