+++ src/actions/images/destroy.ts
@@ -2,7 +2,7 @@
 import prisma from '@/prisma'
 import errorHandler from '@/prisma/errorHandler'
 import type { Image } from '@prisma/client'
-import type { ActionReturn } from '@/actions/type'
+import type { ActionReturn } from '@/actions/Types'
 
 export async function destroyImage(imageId: number): Promise<ActionReturn<Image>> {
     try {
