+++ src/actions/omegaquotes/create.ts
@@ -4,7 +4,7 @@ import prisma from '@/prisma'
 import errorHandler from '@/prisma/errorHandler'
 import { getUser } from '@/auth'
 import { z } from 'zod'
-import type { ActionReturn } from '@/actions/type'
+import type { ActionReturn } from '@/actions/Types'
 import type { OmegaQuote } from '@prisma/client'
 
 export async function createQuote(rawdata: FormData): Promise<ActionReturn<OmegaQuote>> {
