+++ src/actions/omegaquotes/read.ts
@@ -4,7 +4,7 @@ import { readPermissionsOfUser } from '@/actions/permissions/read'
 import prisma from '@/prisma'
 import errorHandler from '@/prisma/errorHandler'
 import { getUser } from '@/auth'
-import type { ActionReturn, ReadPageInput } from '@/actions/type'
+import type { ActionReturn, ReadPageInput } from '@/actions/Types'
 import type { OmegaquoteFiltered } from './Types'
 
 export async function readQuotesPage<const PageSize extends number>(
