+++ src/actions/users/create.ts
@@ -2,7 +2,7 @@
 import prisma from '@/prisma'
 import errorHandler from '@/prisma/errorHandler'
 import { z } from 'zod'
-import type { ActionReturn } from '@/actions/type'
+import type { ActionReturn } from '@/actions/Types'
 import type { User } from '@prisma/client'
 
 export async function createUser(rawdata: FormData): Promise<ActionReturn<User>> {
