+++ src/context/paging/ImageCollectionPaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
 import { readImageCollectionsPage } from '@/actions/images/collections/read'
-import type { ReadPageInput } from '@/actions/type'
+import type { ReadPageInput } from '@/actions/Types'
 import type { ImageCollectionPageReturn } from '@/actions/images/collections/read'
 
 export type PageSizeImageCollection = 12
