+++ src/context/paging/omegaquotesPaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from '@/context/paging/PagingGenerator'
 import { readQuotesPage } from '@/actions/omegaquotes/read'
-import type { ReadPageInput } from '@/actions/type'
+import type { ReadPageInput } from '@/actions/Types'
 import type { OmegaquoteFiltered } from '@/actions/omegaquotes/Types'
 
 export type PageSizeOmegaquote = 20;
